+++ src/Skating.java
@@ -21,6 +21,6 @@ public class Skating extends Activity {
     public Skating clone()
     {return new Skating(this);}
     
-    public double calculateCalories(long duration)
-    {return (double) duration*0.00006;}
+    public int calculateCalories(long duration)
+    {return (int) (duration*0.00006);}
 }
